chore(scheduler): move fcfs into src/scheduler
